Fix download for Edge browser
